refactor: utils.js and tests
